# docs/swagger.yaml
basePath: /
definitions:
  controller.JamKerja:
    properties:
      durasi:
        example: 8
        type: integer
      gmt:
        example: 7
        type: integer
      hari:
        example:
        - Senin
        - Selasa
        - Rabu
        - Kamis
        - Jumat
        - Sabtu
        - Minggu
        items:
          type: string
        type: array
      jam_keluar:
        example: "16:00"
        type: string
      jam_masuk:
        example: "08:00"
        type: string
      piket_tim:
        example: Piket Z
        type: string
      shift:
        example: 2
        type: integer
    type: object
  controller.Karyawan:
    properties:
      hari_kerja:
        example:
        - Senin
        - Selasa
        - Rabu
        - Kamis
        - Jumat
        - Sabtu
        - Minggu
        items:
          type: string
        type: array
      jabatan:
        example: Anonymous
        type: string
      jam_kerja:
        items:
          $ref: '#/definitions/controller.JamKerja'
        type: array
      nama:
        example: Tes Swagger
        type: string
      phone_number:
        example: "08123456789"
        type: string
    type: object
  controller.Presensi:
    properties:
      biodata:
        $ref: '#/definitions/controller.Karyawan'
      checkin:
        description: Datetime     primitive.DateTime `bson:"datetime,omitempty" json:"datetime,omitempty"`
        example: MASUK
        type: string
      latitude:
        example: 123.11
        type: number
      location:
        example: Bandung
        type: string
      longitude:
        example: 123.11
        type: number
      phone_number:
        example: "08123456789"
        type: string
    type: object
host: pemrog3-ulbi.herokuapp.com
info:
  contact:
    email: adecandra1500@gmail.com
    name: API Support
  description: This is a sample swagger for Fiber
  license:
    url: http://github.com/AdeCandra12
  title: TES SWAG
  version: "1.0"
paths:
  /delete/{id}:
    delete:
      consumes:
      - application/json
      description: Hapus data presensi.
      parameters:
      - description: Masukan ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Delete data presensi.
      tags:
      - Presensi
  /ins:
    post:
      consumes:
      - application/json
      description: Input data presensi.
      parameters:
      - description: Payload Body [RAW]
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.Presensi'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Presensi'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Insert data presensi.
      tags:
      - Presensi
  /presensi:
    get:
      consumes:
      - application/json
      description: Mengambil semua data presensi.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Presensi'
      summary: Get All Data Presensi.
      tags:
      - Presensi
  /presensi/{id}:
    get:
      consumes:
      - application/json
      description: Ambil per ID data presensi.
      parameters:
      - description: Masukan ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Presensi'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get By ID Data Presensi.
      tags:
      - Presensi
  /upd/{id}:
    put:
      consumes:
      - application/json
      description: Ubah data presensi.
      parameters:
      - description: Masukan ID
        in: path
        name: id
        required: true
        type: string
      - description: Payload Body [RAW]
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.Presensi'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Presensi'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Update data presensi.
      tags:
      - Presensi
swagger: "2.0"